feat: split views into modules
