+++ packages/workspace-fusion/src/lib/integrations/garden/index.ts
@@ -1,3 +1,2 @@
 export * from './garden';
 export * from './types';
-export * from './components';
