+++ packages/workspace-fusion/src/lib/integrations/power-bi/index.ts
@@ -1,4 +1,2 @@
 export * from './power-bi';
 export * from './types';
-export * from './components';
-export * from './utils';
