--- packages/workspace-fusion/src/lib/integrations/power-bi/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './addPowerBi';
