+++ packages/workspace-fusion/src/modules/garden/components/icons/GardenIcon.tsx
@@ -1,0 +1,33 @@
+export const GardenIcon = (): JSX.Element => {
+	return (
+		<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+			<rect width="24" height="24" />
+
+			<path
+				d="M3 4C2.44772 4 2 4.44772 2 5C2 5.55228 2.44772 6 3 6H21C21.5523 6 22 5.55228 22 5C22 4.44772 21.5523 4 21 4H3Z"
+				fill="#6F6F6F"
+			/>
+
+			<path
+				fillRule="evenodd"
+				clipRule="evenodd"
+				d="M2 9C2 8.44771 2.44772 8 3 8H7C7.55228 8 8 8.44772 8 9V19C8 19.5523 7.55228 20 7 20H3C2.44772 20 2 19.5523 2 19V9ZM4 18V10H6V18H4Z"
+				fill="#6F6F6F"
+			/>
+
+			<path
+				fillRule="evenodd"
+				clipRule="evenodd"
+				d="M9 9C9 8.44772 9.44771 8 10 8H14C14.5523 8 15 8.44772 15 9V15C15 15.5523 14.5523 16 14 16H10C9.44771 16 9 15.5523 9 15V9ZM11 14V10H13V14H11Z"
+				fill="#6F6F6F"
+			/>
+
+			<path
+				fillRule="evenodd"
+				clipRule="evenodd"
+				d="M16 9C16 8.44772 16.4477 8 17 8H21C21.5523 8 22 8.44772 22 9V15C22 15.5523 21.5523 16 21 16H17C16.4477 16 16 15.5523 16 15V9ZM18 14V10H20V14H18Z"
+				fill="#6F6F6F"
+			/>
+		</svg>
+	);
+};
