+++ packages/workspace-fusion/src/modules/garden/components/index.ts
@@ -1,3 +1,2 @@
 export * from './workspace-header';
 export * from './wrapper';
-export * from './addGarden';
