+++ packages/workspace-fusion/src/modules/garden/icons/GardenIcon.tsx
