+++ packages/workspace-fusion/src/modules/power-bi/icons/PowerBiIcon.tsx
